+++ packages/koop-cache-memory/test/index.js
@@ -2,29 +2,41 @@ const test = require('tape')
 const Cache = require('../src')
 const cache = new Cache()
 const _ = require('lodash')
-const geojson = {
-  type: 'FeatureCollection',
-  metadata: {
-    name: 'Test',
-    description: 'Test'
-  },
-  features: [
-    {
-      type: 'Feature',
+const { asCachableGeojson } = require('../src/helper')
+
+function getFeatures() {
+  return {
+    type: 'FeatureCollection',
+    crs: {
+      type: 'name',
       properties: {
-        key: 'value'
-      },
-      geometry: {
-        foo: 'bar'
+        type: 'EPSG:4326'
       }
-    }
-  ]
+    },
+    metadata: {
+      name: 'Test',
+      description: 'Test'
+    },
+    features: [
+      {
+        type: 'Feature',
+        properties: {
+          key: 'value'
+        },
+        geometry: {
+          foo: 'bar'
+        }
+      }
+    ]
+  }
 }
 
 test('Inserting and retreiving from the cache', t => {
+  const geojson = getFeatures();
   cache.insert('key', geojson, {ttl: 600})
   const cached = cache.retrieve('key')
   t.equal(cached.features[0].properties.key, 'value', 'retrieved features')
+  t.equal(cached.crs.type, 'name', 'retrieved coordinate reference system')
   t.equal(cached.metadata.name, 'Test', 'retrieved metadata')
   t.ok(cached.metadata.expires, 'expiration set')
   t.ok(cached.metadata.updated, 'updated set')
@@ -32,6 +44,7 @@ test('Inserting and retreiving from the cache', t => {
 })
 
 test('Upserting and streaming from the cache', t => {
+  const geojson = getFeatures();
   cache.upsert('key', geojson, {ttl: 600})
   const readstream = cache.createStream('key')
   readstream.on("data", (chunk) => {
@@ -41,9 +54,11 @@ test('Upserting and streaming from the cache', t => {
 })
 
 test('Inserting and retreiving from the cache using upsert when the cache is empty', t => {
+  const geojson = getFeatures();
   cache.upsert('keyupsert', geojson, {ttl: 600})
   const cached = cache.retrieve('keyupsert')
   t.equal(cached.features[0].properties.key, 'value', 'retrieved features')
+  t.equal(cached.crs.type, 'name', 'retrieved coordinate reference system')
   t.equal(cached.metadata.name, 'Test', 'retrieved metadata')
   t.ok(cached.metadata.expires, 'expiration set')
   t.ok(cached.metadata.updated, 'updated set')
@@ -51,12 +66,14 @@ test('Inserting and retreiving from the cache using upsert when the cache is emp
 })
 
 test('Inserting and retreiving from the cache using upsert when the cache is filled', t => {
+  const geojson = getFeatures();
   cache.insert('keyupsertupdate', geojson, {ttl: 600})
   const geojson2 = _.cloneDeep(geojson)
   geojson2.features[0].properties['key'] = 'updated'
   cache.upsert('keyupsertupdate', geojson2, {ttl: 600})
   const cached = cache.retrieve('keyupsertupdate')
   t.equal(cached.features[0].properties.key, 'updated', 'retrieved features')
+  t.equal(cached.crs.type, 'name', 'retrieved coordinate reference system')
   t.equal(cached.metadata.name, 'Test', 'retrieved metadata')
   t.ok(cached.metadata.expires, 'expiration set')
   t.ok(cached.metadata.updated, 'updated set')
@@ -64,11 +81,13 @@ test('Inserting and retreiving from the cache using upsert when the cache is fil
 })
 
 test('Inserting and retreiving from the cache callback style', t => {
+  const geojson = getFeatures();
   cache.insert('keyb', geojson, {ttl: 600}, (err) => {
     t.error(err, 'no error')
     const cached = cache.retrieve('keyb')
-    t.equal(cached.features[0].properties.key, 'value', 'retrieved features')
-    t.equal(cached.metadata.name, 'Test', 'retrieved metadata')
+  t.equal(cached.features[0].properties.key, 'value', 'retrieved features')
+  t.equal(cached.crs.type, 'name', 'retrieved coordinate reference system')
+  t.equal(cached.metadata.name, 'Test', 'retrieved metadata')
     t.ok(cached.metadata.expires, 'expiration set')
     t.ok(cached.metadata.updated, 'updated set')
     t.end()
@@ -76,10 +95,12 @@ test('Inserting and retreiving from the cache callback style', t => {
 })
 
 test('Inserting and appending to the cache', t => {
+  const geojson = getFeatures();
   cache.insert('key2', geojson, {ttl: 600})
   cache.append('key2', geojson)
   const cached = cache.retrieve('key2')
   t.equal(cached.features.length, 2, 'retrieved all features')
+  t.equal(cached.crs.type, 'name', 'retrieved coordinate reference system')
   t.equal(cached.metadata.name, 'Test', 'retrieved metadata')
   t.ok(cached.metadata.expires, 'expiration set')
   t.ok(cached.metadata.updated, 'updated set')
@@ -87,6 +108,7 @@ test('Inserting and appending to the cache', t => {
 })
 
 test('Updating an existing entry in the cache', t => {
+  const geojson = getFeatures();
   cache.insert('key3', geojson, {ttl: 600})
   const geojson2 = _.cloneDeep(geojson)
   geojson2.features[0].properties.key = 'test2'
@@ -94,6 +116,7 @@ test('Updating an existing entry in the cache', t => {
   const cached = cache.retrieve('key3')
   t.equal(cached.features[0].properties.key, 'test2', 'retrieved only new features')
   t.equal(cached.features.length, 1, 'retrieved only new features')
+  t.equal(cached.crs.type, 'name', 'retrieved original coordinate reference system')
   t.equal(cached.metadata.name, 'Test', 'retrieved original metadata')
   t.ok(cached.metadata.expires, 'expiration set')
   t.ok(cached.metadata.updated, 'updated set')
@@ -101,6 +124,7 @@ test('Updating an existing entry in the cache', t => {
 })
 
 test('Inserting and deleting from the cache', t => {
+  const geojson = getFeatures();
   t.plan(2)
   cache.insert('key4', geojson)
   cache.delete('key4')
@@ -111,6 +135,7 @@ test('Inserting and deleting from the cache', t => {
 })
 
 test('Trying to call insert when something is already in the cache', t => {
+  const geojson = getFeatures();
   t.plan(2)
   cache.insert('key5', geojson)
   cache.insert('key5', geojson, {}, err => {
@@ -120,10 +145,28 @@ test('Trying to call insert when something is already in the cache', t => {
 })
 
 test('Trying to delete the catalog entry when something is still in the cache', t => {
+  const geojson = getFeatures();
   t.plan(2)
   cache.insert('key6', geojson)
   cache.catalog.delete('key6', err => {
     t.ok(err, 'Should return an error')
     t.equal(err.message, 'Cannot delete catalog entry while data is still in cache', 'Error should have correct message')
   })
+})
+
+test('Helper prepares geojson for cache', t => {
+  const full = getFeatures()
+  const features = full.features
+  const empty = {}
+  const fullGeojson = asCachableGeojson(full)
+  const featuresAsGeojson = asCachableGeojson(features)
+  const emptyAsGeojson = asCachableGeojson(empty)
+  const nullAsGeojson = asCachableGeojson(null)
+  const undefinedAsGeojson = asCachableGeojson(undefined)
+  t.equal(fullGeojson.features[0].properties.key, 'value', 'Full geojson stays geojson')  
+  t.equal(featuresAsGeojson.features[0].properties.key, 'value', 'Features is converted to geojson') 
+  t.equal(emptyAsGeojson.features.length, 0, 'Empty object becomes geojson')
+  t.equal(nullAsGeojson.features.length, 0, 'Null object becomes geojson')
+  t.equal(undefinedAsGeojson.features.length, 0, 'Undefined object becomes geojson')
+  t.end()
 })